Avoid buffer overflow in `runStreamingBuilder`
In `runStreamingBuilder`, if the `Builder` results in `More`, we were discarding
the minimum required buffer size for the continuation contained in the `More`.
In some (apparently rare) circumstances, this minimum required size may be less
than the actual space we have left in the buffer, resulting in a buffer overflow
when we run the writer. We now compare the room left in the buffer with the
minimum size indicated by the `More` and flush before writing instead of
overflowing the buffer in such cases.