Introduce `responseStreamingIface`
Prior to this commit, when a server finished streaming, it had no way of
indicating this (i.e. it has no way of setting the second argument to
`StreamingBuilder`). As a result, `http2` would emit a final DATA frame when it
finished the processing the builder along with a subsequent *empty* data frame
when processing `StreamingFinished`. This results in an empty data frame at the
end of every response. If the volume of responses is high, this even triggers
the `emptyFrameRateLimit` in `http2`.